Permalinks
Bugs that prevent new code from being saved to the SQL table have been
found and most of them squashed.  You can now generate permalinks and
load projects and edit them with just the permalinks.  When loading text
from the SQL table ACE interprets new lines as \n; the result is one
line of code with a large number of \ns.  This needs to be fixed.